Fix for global variables to work correctly in TweakMode.
The previous code that handled global variables would only work if the
global variables were placed above all other functions in the sketch,
i.e. before setup() and draw(). If global variables were placed anywhere
else in the program order, they would be treated as non-globals.

This causes problems, as the global values affected will not be loaded &
accessible before the rest of the sketch logic.